Move files from setup module to brand new latest module
