+++ app/k9mail/build.gradle
@@ -7,7 +7,8 @@ if (rootProject.testCoverage) {
 
 dependencies {
     implementation project(":app:ui:legacy")
-    implementation project(":app:ui:setup")
+    debugImplementation project(":app:ui:setup")
+    implementation project(":app:ui:latest")
     implementation project(":app:core")
     implementation project(":app:storage")
     implementation project(":app:crypto-openpgp")
