+++ app/ui/latest/build.gradle
@@ -7,13 +7,18 @@ dependencies {
     implementation project(":app:autodiscovery:api")
     implementation project(":mail:common")
 
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:${versions.androidxLifecycle}"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:${versions.androidxLifecycle}"
-    implementation "androidx.constraintlayout:constraintlayout:${versions.androidxConstraintLayout}"
     implementation "androidx.core:core-ktx:${versions.androidxCore}"
     implementation "com.jakewharton.timber:timber:${versions.timber}"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:${versions.kotlinCoroutines}"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:${versions.kotlinCoroutines}"
+
+    implementation "androidx.activity:activity-compose:${versions.compose}"
+    implementation "androidx.compose.ui:ui-tooling:${versions.compose}"
+    implementation "androidx.compose.ui:ui:${versions.compose}"
+    implementation "androidx.compose.foundation:foundation:${versions.compose}"
+    implementation "androidx.compose.material:material:${versions.compose}"
+    implementation "androidx.navigation:navigation-compose:${versions.composeNavigation}"
+    implementation "io.insert-koin:koin-androidx-compose:${versions.koin}"
 
     testImplementation project(':mail:testing')
     testImplementation project(':app:testing')
@@ -37,6 +42,14 @@ android {
     lintOptions {
         abortOnError false
         lintConfig file("$rootProject.projectDir/config/lint/lint.xml")
+    }
+
+    buildFeatures {
+        compose true
+    }
+
+    composeOptions {
+        kotlinCompilerExtensionVersion = versions.compose
     }
 
     compileOptions {
