+++ app/ui/latest/src/main/java/com/mudita/mail/ComposeUiModules.kt
@@ -1,0 +1,9 @@
+package com.mudita.mail
+
+import com.mudita.mail.ui.di.composeNavigationModule
+import com.mudita.mail.ui.di.composeViewModelModule
+
+val composeUiModules = listOf(
+    composeNavigationModule,
+    composeViewModelModule
+)
