+++ app/ui/latest/src/main/java/com/mudita/mail/ui/di/ViewModelModule.kt
@@ -1,0 +1,14 @@
+package com.mudita.mail.ui.di
+
+import com.mudita.mail.ui.extension.getFrom
+import com.mudita.mail.ui.usecase.signIn.viewModel.SignInViewModel
+import org.koin.dsl.module
+
+val composeViewModelModule = module {
+
+    factory { params ->
+        SignInViewModel(
+            navigator = getFrom(params)
+        )
+    }
+}
