+++ app/ui/latest/src/main/java/com/mudita/mail/ui/extension/GetFrom.kt
@@ -1,0 +1,9 @@
+package com.mudita.mail.ui.extension
+
+import org.koin.core.parameter.ParametersHolder
+import org.koin.core.parameter.parametersOf
+import org.koin.core.scope.Scope
+
+inline fun <reified T : Any> Scope.getFrom(
+    params: ParametersHolder
+) = get<T> { parametersOf(*params.values.toTypedArray()) }
