+++ app/ui/latest/src/main/java/com/mudita/mail/ui/navigation/Destination.kt
@@ -1,0 +1,20 @@
+package com.mudita.mail.ui.navigation
+
+object SignInDestination : Destination("signIn")
+
+sealed class Destination(
+    protected open val route: String,
+) {
+
+    open fun toRoute() = route
+}
+
+sealed class Parametrized<T : Any>(
+    override val route: String,
+    val key: String,
+) : Destination(route) {
+
+    final override fun toRoute() = "$route/{$key}"
+
+    fun toRoute(param: T) = "$route/$param"
+}
