+++ app/ui/latest/src/main/java/com/mudita/mail/ui/navigation/navigator/Navigator.kt
@@ -1,0 +1,3 @@
+package com.mudita.mail.ui.navigation.navigator
+
+interface Navigator
