+++ app/ui/latest/src/main/java/com/mudita/mail/ui/theme/Colors.kt
@@ -1,0 +1,10 @@
+package com.mudita.mail.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val PaleRed = Color(0xffd85252)
+val Hint = Color(0xFF8A94A6)
+val WhitePure = Color.White
+val GreyLight = Color(0xFFEBEBEB)
+val BlackPure = Color.Black
+val GreyDark = Color(0xFF545A5E)
