+++ app/ui/latest/src/main/java/com/mudita/mail/ui/theme/Type.kt
@@ -1,0 +1,48 @@
+package com.mudita.mail.ui.theme
+
+import androidx.compose.material.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+
+val MuditaTypography = Typography(
+    h4 = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 32.sp,
+    ),
+    h5 = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 24.sp,
+    ),
+    h6 = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 18.sp
+    ),
+    subtitle1 = TextStyle(
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp
+    ),
+    subtitle2 = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 14.sp
+    ),
+    body1 = TextStyle(
+        fontWeight = FontWeight.Normal,
+        fontSize = 14.sp
+    ),
+    body2 = TextStyle(
+        fontSize = 14.sp,
+    ),
+    button = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 18.sp
+    ),
+    caption = TextStyle(
+        fontWeight = FontWeight.Normal,
+        fontSize = 12.sp
+    ),
+    overline = TextStyle(
+        fontWeight = FontWeight.Medium,
+        fontSize = 12.sp
+    )
+)
