+++ app/ui/latest/src/main/java/com/mudita/mail/ui/usecase/signIn/navigator/SignInNavigator.kt
@@ -1,0 +1,8 @@
+package com.mudita.mail.ui.usecase.signIn.navigator
+
+import com.mudita.mail.ui.navigation.navigator.Navigator
+
+interface SignInNavigator : Navigator {
+
+    fun moveToCredentials()
+}
