+++ app/ui/latest/src/main/java/com/mudita/mail/ui/usecase/signIn/navigator/SignInScreenNavigatorImpl.kt
@@ -1,0 +1,12 @@
+package com.mudita.mail.ui.usecase.signIn.navigator
+
+import android.content.Context
+import com.fsck.k9.ui.base.navigation.ToSetupAccountNavigator
+
+class SignInScreenNavigatorImpl(
+    private val context: Context,
+    private val toSetupAccountNavigator: ToSetupAccountNavigator
+) : SignInNavigator {
+
+    override fun moveToCredentials() = toSetupAccountNavigator.moveToSetupAccount(context)
+}
