--- app/ui/setup/src/main/java/com/fsck/k9/ui/setup/ComposeUiModules.kt
@@ -1,9 +1,0 @@
-package com.fsck.k9.ui.setup
-
-import com.fsck.k9.ui.setup.ui.di.composeNavigationModule
-import com.fsck.k9.ui.setup.ui.di.composeViewModelModule
-
-val composeUiModules = listOf(
-    composeNavigationModule,
-    composeViewModelModule
-)
