--- app/ui/setup/src/main/java/com/fsck/k9/ui/setup/ui/navigation/MuditaNavGraph.kt
@@ -1,30 +1,0 @@
-package com.fsck.k9.ui.setup.ui.navigation
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.platform.LocalContext
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.fsck.k9.ui.setup.ui.usecase.signIn.view.SignInScreen
-import org.koin.androidx.compose.getViewModel
-import org.koin.core.parameter.parametersOf
-
-@Composable
-fun MuditaMailNavGraph(
-    navController: NavHostController = rememberNavController()
-) {
-    NavHost(
-        navController = navController,
-        startDestination = SignInDestination.toRoute(),
-    ) {
-        composable(SignInDestination.toRoute()) {
-            val context = LocalContext.current
-            SignInScreen(
-                getViewModel {
-                    parametersOf(context)
-                }
-            )
-        }
-    }
-}
