--- app/ui/setup/src/main/java/com/fsck/k9/ui/setup/ui/navigation/navigator/Navigator.kt
@@ -1,3 +1,0 @@
-package com.fsck.k9.ui.setup.ui.navigation.navigator
-
-interface Navigator
