--- app/ui/setup/src/main/java/com/fsck/k9/ui/setup/ui/theme/Colors.kt
@@ -1,10 +1,0 @@
-package com.fsck.k9.ui.setup.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val PaleRed = Color(0xffd85252)
-val Hint = Color(0xFF8A94A6)
-val WhitePure = Color.White
-val GreyLight = Color(0xFFEBEBEB)
-val BlackPure = Color.Black
-val GreyDark = Color(0xFF545A5E)
