--- app/ui/setup/src/main/java/com/fsck/k9/ui/setup/ui/usecase/signIn/navigator/SignInNavigator.kt
@@ -1,8 +1,0 @@
-package com.fsck.k9.ui.setup.ui.usecase.signIn.navigator
-
-import com.fsck.k9.ui.setup.ui.navigation.navigator.Navigator
-
-interface SignInNavigator : Navigator {
-
-    fun moveToCredentials()
-}
