+++ settings.gradle
@@ -24,3 +24,4 @@ include ':backend:webdav'
 include ':backend:jmap'
 include ':backend:demo'
 include ':plugins:openpgp-api-lib:openpgp-api'
+include ':app:ui:latest'
